pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "o2-cli"
version = "0.1.0"
description = "CLI for O2 DEX Trading Platform"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "respx>=0.20.0",
    "ruff>=0.1.0",
]

[project.scripts]
o2 = "o2_cli.cli:app"

[tool.setuptools.packages.find]
include = ["o2_cli*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
